+++ src/main/java/in/gov/kfon/dmdm/contract/StateResponse.java
@@ -1,18 +1,17 @@
 package in.gov.kfon.dmdm.contract;
 
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class StateResponse {
-    private UUID id;
-    private String code;
-    private String name;
+  private UUID id;
+  private String code;
+  private String name;
 }