+++ src/test/java/in/gov/kfon/dmdm/controller/PincodeControllerTest.java
@@ -133,6 +133,7 @@ class PincodeControllerTest {
   void testFetchPostOfficeDetailsByPincode() throws Exception {
 
     Integer pincode = 695001;
+    UUID districtId = UUID.randomUUID();
 
     CommonLookUp lookup = new CommonLookUp();
     lookup.setId(id);
@@ -141,25 +142,24 @@ class PincodeControllerTest {
     lookup.setNameInLocal("തിരുവനന്തപുരം ജിപിഒ");
     lookup.setIsActive(true);
     lookup.setDistrict("Thiruvananthapuram");
-    lookup.setDistrictId(1);
+    lookup.setDistrictId(districtId);
 
     List<CommonLookUp> list = List.of(lookup);
 
     when(service.fetchPostOfficeDetailsByPincode(pincode)).thenReturn(list);
 
     mockMvc
-            .perform(get("/api/pincode/post-office-details/{pincode}", pincode))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.message").value("Fetched post offices by pincode"))
-            .andExpect(jsonPath("$.data").isArray())
-            .andExpect(jsonPath("$.data.length()").value(1))
-            .andExpect(jsonPath("$.data[0].id").value(id.toString()))
-            .andExpect(jsonPath("$.data[0].code").value("695001"))
-            .andExpect(jsonPath("$.data[0].name").value("Thiruvananthapuram GPO"))
-            .andExpect(jsonPath("$.data[0].nameInLocal").value("തിരുവനന്തപുരം ജിപിഒ"))
-            .andExpect(jsonPath("$.data[0].district").value("Thiruvananthapuram"))
-            .andExpect(jsonPath("$.data[0].districtId").value(1))
-            .andExpect(jsonPath("$.data[0].isActive").value(true));
+        .perform(get("/api/pincode/post-office-details/{pincode}", pincode))
+        .andExpect(status().isOk())
+        .andExpect(jsonPath("$.message").value("Fetched post offices by pincode"))
+        .andExpect(jsonPath("$.data").isArray())
+        .andExpect(jsonPath("$.data.length()").value(1))
+        .andExpect(jsonPath("$.data[0].id").value(id.toString()))
+        .andExpect(jsonPath("$.data[0].code").value("695001"))
+        .andExpect(jsonPath("$.data[0].name").value("Thiruvananthapuram GPO"))
+        .andExpect(jsonPath("$.data[0].nameInLocal").value("തിരുവനന്തപുരം ജിപിഒ"))
+        .andExpect(jsonPath("$.data[0].district").value("Thiruvananthapuram"))
+        .andExpect(jsonPath("$.data[0].districtId").value(districtId.toString()))
+        .andExpect(jsonPath("$.data[0].isActive").value(true));
   }
-
 }